Android: IBiometricKeyEncryptor should not require Android "M" - biometry package hides version dependency in public ifcs Android: IAddBiometryFactorListener moved to "biometry" package - Right exception returned in error callback
